LC-count-total-number-of-colored-cells [Runtime: N/A; Memory: N/A]

+++ math/count-total-number-of-colored-cells.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long coloredCells(int n) {
-        int ans = 1;
+        long ans = 1;
         for(int i=0; i<n; i++){
             ans=ans+(1*(4*i));
         }
